+++ src/vscode/treeviews/clusters/DatabricksCluster.ts
@@ -235,8 +235,10 @@ export class DatabricksCluster extends DatabricksClusterTreeItem {
 		setTimeout(() => vscode.commands.executeCommand("databricksClusters.refresh", undefined, false), 1000);
 	}
 
-	async showDefinition(): Promise<vscode.TextDocument> {
-		return vscode.workspace.openTextDocument({ language: "json", content: JSON.stringify(this.definition, null, "\t") });
+	async showDefinition(): Promise<void> {
+		vscode.workspace.openTextDocument({ language: "json", content: JSON.stringify(this.definition, null, "\t") }).then(
+			document => vscode.window.showTextDocument(document)
+		);
 	}
 
 	async useForSQL(): Promise<void> {